Move GraphQLTestClient in tests
