--- strawberry/django/test/__init__.py
@@ -1,4 +1,0 @@
-from .client import GraphQLTestClient
-
-
-__all__ = ["GraphQLTestClient"]
