--- strawberry/test/__init__.py
@@ -1,4 +1,0 @@
-from .client import BaseGraphQLTestClient, Body, Response
-
-
-__all__ = ["BaseGraphQLTestClient", "Body", "Response"]
