+++ tests/django/django_settings.py
@@ -1,6 +1,7 @@
 SECRET_KEY = 1
 
 INSTALLED_APPS = ["strawberry.django", "tests.django.app"]
+ROOT_URLCONF = "tests.django.app.urls"
 
 TEMPLATES = [
     {
